# openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "ccfatigue",
    "version": "0.1.0"
  },
  "paths": {
...
    "/analysis/cycleCounting/file": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Run Cycle Counting File",
        "operationId": "run_cycle_counting_file",
        "parameters": [
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CycleCountingMethod"
            },
            "name": "method",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_cycle_counting_file_analysis_cycleCounting_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Cycle Counting File Analysis Cyclecounting File Post",
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/snCurve/file": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Run Sn Curve File",
        "operationId": "run_sn_curve_file",
        "parameters": [
          {
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/SnCurveMethod"
              }
            },
            "name": "methods",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Rratios",
              "type": "array",
              "items": {
                "type": "number"
              }
            },
            "name": "rRatios",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_sn_curve_file_analysis_snCurve_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnCurveResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/cld/file": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Run Cld File",
        "operationId": "run_cld_file",
        "parameters": [
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CldMethod"
            },
            "name": "method",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_cld_file_analysis_cld_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Cld File Analysis Cld File Post",
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/fatigueFailure/file": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Run Fatigue Failure File",
        "operationId": "run_fatigue_failure_file",
        "parameters": [
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FatigueFailureMethod"
            },
            "name": "method",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SnModel"
            },
            "name": "snModel",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Desirableangle",
              "type": "number"
            },
            "name": "desirableAngle",
            "in": "query"
          },
          {
            "required": true,
            "schema": {
              "title": "Offaxisangle",
              "type": "number"
            },
            "name": "offAxisAngle",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_fatigue_failure_file_analysis_fatigueFailure_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Fatigue Failure File Analysis Fatiguefailure File Post",
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/analysis/damageSummation/file": {
      "post": {
        "tags": [
          "analysis"
        ],
        "summary": "Run Damage Summation File",
        "operationId": "run_damage_summation_file",
        "parameters": [
          {
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/DamageSummationMethod"
            },
            "name": "method",
            "in": "query"
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_run_damage_summation_file_analysis_damageSummation_file_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Run Damage Summation File Analysis Damagesummation File Post",
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Body_run_cld_file_analysis_cld_file_post": {
        "title": "Body_run_cld_file_analysis_cld_file_post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Body_run_cycle_counting_file_analysis_cycleCounting_file_post": {
        "title": "Body_run_cycle_counting_file_analysis_cycleCounting_file_post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Body_run_damage_summation_file_analysis_damageSummation_file_post": {
        "title": "Body_run_damage_summation_file_analysis_damageSummation_file_post",
        "required": [
          "sncFile",
          "cycFile"
        ],
        "type": "object",
        "properties": {
          "sncFile": {
            "title": "Sncfile",
            "type": "string",
            "format": "binary"
          },
          "cycFile": {
            "title": "Cycfile",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Body_run_fatigue_failure_file_analysis_fatigueFailure_file_post": {
        "title": "Body_run_fatigue_failure_file_analysis_fatigueFailure_file_post",
        "required": [
          "xFile",
          "yFile",
          "fFile"
        ],
        "type": "object",
        "properties": {
          "xFile": {
            "title": "Xfile",
            "type": "string",
            "format": "binary"
          },
          "yFile": {
            "title": "Yfile",
            "type": "string",
            "format": "binary"
          },
          "fFile": {
            "title": "Ffile",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "Body_run_sn_curve_file_analysis_snCurve_file_post": {
        "title": "Body_run_sn_curve_file_analysis_snCurve_file_post",
        "required": [
          "file"
        ],
        "type": "object",
        "properties": {
          "file": {
            "title": "File",
            "type": "string",
            "format": "binary"
          }
        }
      },
      "CldMethod": {
        "title": "CldMethod",
        "enum": [
          "Harris"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "CycleCountingMethod": {
        "title": "CycleCountingMethod",
        "enum": [
          "RangeMean"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "DamageSummationMethod": {
        "title": "DamageSummationMethod",
        "enum": [
          "Harris"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "EchartLine": {
        "title": "EchartLine",
        "required": [
          "name",
          "xData",
          "yData"
        ],
        "type": "object",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "xData": {
            "title": "Xdata",
            "type": "array",
            "items": {
              "type": "number"
            }
          },
          "yData": {
            "title": "Ydata",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        }
      },
...
      "FatigueFailureMethod": {
        "title": "FatigueFailureMethod",
        "enum": [
          "FTPT"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "HTTPValidationError": {
        "title": "HTTPValidationError",
        "type": "object",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
...
      "SnCurveMethod": {
        "title": "SnCurveMethod",
        "enum": [
          "LinLog",
          "LogLog",
          "Sendeckyj",
          "Whitney"
        ],
        "type": "string",
        "description": "An enumeration."
      },
      "SnCurveResult": {
        "title": "SnCurveResult",
        "required": [
          "outputs",
          "lines"
        ],
        "type": "object",
        "properties": {
          "outputs": {
            "title": "Outputs",
            "type": "object",
            "additionalProperties": {
              "type": "string",
              "format": "binary"
            }
          },
          "lines": {
            "title": "Lines",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EchartLine"
            }
          }
        }
      },
      "SnModel": {
        "title": "SnModel",
        "enum": [
          "Lin-Log",
          "Log-Log"
        ],
        "type": "string",
        "description": "An enumeration."
      },
...
      "ValidationError": {
        "title": "ValidationError",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "type": "object",
        "properties": {
          "loc": {
            "title": "Location",
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            }
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        }
      }
    }
  }
}